Replace "which" with "command -v".
Split2flac is incompatible with the dash (a POSIX compliant /bin/sh
implementation) due to the use of the "&>" output redirection operator in
conjunction with the "which" commands.  In order to fix this, the "&>" operator
has been replaced by ">", and (in order to avoid the outputs to stderr) "which"
has been replaced by "command -v".